research long long

+++ Sources/CvOutcomeList.cpp
@@ -169,7 +169,7 @@ bool CvOutcomeList::execute(CvUnit &kUnit, PlayerTypes eDefeatedUnitPlayer, Unit
 	return false;
 }
 
-int CvOutcomeList::AI_getValueInPlot(const CvUnit& kUnit, const CvPlot& kPlot, bool bForTrade) const
+int64_t CvOutcomeList::AI_getValueInPlot(const CvUnit& kUnit, const CvPlot& kPlot, bool bForTrade) const
 {
 	std::vector<std::pair<CvOutcome*, int> > apOutcome;
 	std::set<OutcomeTypes> aeReplacedOutcomes;
@@ -206,7 +206,7 @@ int CvOutcomeList::AI_getValueInPlot(const CvUnit& kUnit, const CvPlot& kPlot, b
 		iChanceSum = 100;
 	}
 
-	int iValue = 0;
+	int64_t iValue = 0;
 
 	for (int i=0; i<(int)apOutcome.size(); i++)
 	{
