+++ Sources/CvStructs.h
@@ -530,7 +530,7 @@ struct MissionData
 struct TradeData
 {
 	TradeableItems m_eItemType; //What type of item is this
-	int m_iData; //Any additional data?
+	int64_t m_iData; //Any additional data?
 	bool m_bOffering; // Is this item up for grabs?
 	bool m_bHidden; // Are we hidden?
