+++ Sources/CyPlayer.cpp
@@ -681,7 +681,7 @@ int CyPlayer::getAdvancedStartRouteCost(int /*RouteTypes*/ eRoute, bool bAdd, co
 	return m_pPlayer->getAdvancedStartRouteCost((RouteTypes) eRoute, bAdd, NULL != pPlot ? pPlot->getPlot() : NULL);
 }
 
-int CyPlayer::getAdvancedStartTechCost(int /*TechTypes*/ eTech, bool bAdd) const
+int64_t CyPlayer::getAdvancedStartTechCost(int /*TechTypes*/ eTech, bool bAdd) const
 {
 	return m_pPlayer->getAdvancedStartTechCost((TechTypes) eTech, bAdd);
 }
@@ -836,7 +836,7 @@ int CyPlayer::getNumMilitaryUnits() const
 	return m_pPlayer->getNumMilitaryUnits();
 }
 
-int CyPlayer::getOverflowResearch() const
+uint64_t CyPlayer::getOverflowResearch() const
 {
 	return m_pPlayer->getOverflowResearch();
 }
